Renamed unittest project and removed libman from build step
