+++ Empleado.sln
@@ -10,9 +10,7 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Data", "Data\Data.csproj", 
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Migrations", "Migrations\Migrations.csproj", "{AAED3C8C-0CD2-4199-8EBA-5BEFD09EDA32}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Test", "Test", "{34E69831-9280-4D3B-ACA0-5D0A09AF7B85}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Empleado.Unit.Tests", "Tests\Empleado.Unit.Tests\Empleado.Unit.Tests.csproj", "{23703171-B45F-4BC7-BEE0-B4C2FF999EBC}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "UnitTests", "UnitTests\UnitTests.csproj", "{D0F022A1-860D-41BA-8620-E70637843022}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -40,9 +38,9 @@ Global
 		{AAED3C8C-0CD2-4199-8EBA-5BEFD09EDA32}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{AAED3C8C-0CD2-4199-8EBA-5BEFD09EDA32}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{AAED3C8C-0CD2-4199-8EBA-5BEFD09EDA32}.Release|Any CPU.Build.0 = Release|Any CPU
-		{23703171-B45F-4BC7-BEE0-B4C2FF999EBC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{23703171-B45F-4BC7-BEE0-B4C2FF999EBC}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{23703171-B45F-4BC7-BEE0-B4C2FF999EBC}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{23703171-B45F-4BC7-BEE0-B4C2FF999EBC}.Release|Any CPU.Build.0 = Release|Any CPU	
+		{D0F022A1-860D-41BA-8620-E70637843022}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D0F022A1-860D-41BA-8620-E70637843022}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D0F022A1-860D-41BA-8620-E70637843022}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D0F022A1-860D-41BA-8620-E70637843022}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal
