+++ UnitTests/Framework/BaseTest.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using NUnit.Framework;
+
+namespace UnitTests.Framework
+{
+    public abstract class BaseTest<T>
+    {
+        protected T _sut;
+
+        [SetUp]
+        public abstract void Setup();
+
+    }
+}
